TDS-Solver-master/requirements.txt
Flask>=2.0.0
requests>=2.0.0
python-dotenv>=0.21.0
pytest>=7.0.0